functions/requirements.txt
firebase_functions~=0.1.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
firebase-admin==6.2.0
python-jose==3.3.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
transformers==4.37.2
torch==2.2.0
pytest==7.4.3
pytest-asyncio==0.23.5
httpx==0.25.1
passlib==1.7.4